src/requirements.txt
numpy==1.21.0
pyfaidx==0.5.9.5
pysam==0.16.0.1
biopython==1.79
